requirements.txt
streamlit
Pillow
tensorflow
numpy
pandas
opencv-python
matplotlib
scikit-learn
